Move datafile.
