+++ tests/test_exact_cover.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from exact_cover import get_exact_cover
-from .polyomino_data import polyomino_problem
+from .helpers.polyomino_data import polyomino_problem
 
 
 def test_exact_cover():
